refactor: 💡 move tooltip as a separate component
